[AutoTimer] Allow autotimers on individual IPTV streams
AutoTimer.py: Add handling to process individual IPTV streams.

AutoTimerImporter.py and AutoTimerEditor.py: Restrict GUI so autotimers set from an IPTV stream are restricted to that stream only.

Hopefully this is a good compromise of functionality and resources. The use of resources remains low and the receiver continues to be responsive without continual spinners encountered previously.